feature/send-emails-issue-63: try to use resend to send contact form
