+++ package.json
@@ -38,6 +38,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-email": "1.9.1",
+    "resend": "0.12.2",
     "sharp": "0.31.3"
   },
   "devDependencies": {
